+++ src/modules/client/services/DeleteClientService.ts
@@ -1,6 +1,7 @@
 import { Client } from '@prisma/client';
 import IClientsRepository from '../domain/repositories/IClientsRepository';
 import ClientsRepository from '../infra/repositories/prisma/ClientsRepository';
+import AppError from '@shared/errors/AppError';
 
 export default class DeleteClientService {
   private clientsRepository: IClientsRepository;
@@ -9,10 +10,20 @@ export default class DeleteClientService {
     this.clientsRepository = new ClientsRepository();
   }
 
-  public async execute(clientId: number): Promise<Client> {
+  public async execute(clientId: number): Promise<Client | AppError> {
+    try {
+      const client = await this.clientsRepository.findById(clientId);
 
-    const deletedClient = await this.clientsRepository.delete(clientId);
+      if (!client) throw new AppError('Cliente não encontrado.');
 
-    return deletedClient;
+      const deletedClient = await this.clientsRepository.delete(clientId);
+
+      return deletedClient;
+    } catch (error) {
+      if (error instanceof AppError) return error;
+      console.error(error);
+
+      return new AppError(`${error}`);
+    }
   }
 }
